move the implementation of `From` to mod.rs
